+++ Changelog.md
@@ -9,6 +9,7 @@
 - Allow instructors to configure an end date until which students can run automated tests (#6992)
 - Support Jupyter notebooks for results printing (#6993)
 - Gave TAs read-only access to starter file information under assignment settings (#6996)
+- Enable bulk download of print PDFs for an assignments (#6998)
 - Allow inactive groups in the submissions table to be toggled for display (#7000)
 - Fixed long annotations being cut off in the annotation table (#7001)